scripts/upscaler/requirements.txt
# Real-ESRGAN Training and Inference Dependencies
# Install: pip install -r requirements.txt

# Core ML
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.20,<2.0

# Real-ESRGAN
basicsr>=1.4.2
realesrgan>=0.3.0

# Image processing
opencv-python>=4.5
pillow>=9.0

# Data
requests>=2.28
tqdm>=4.64
lmdb>=1.4.0

# Evaluation metrics
lpips>=0.1.4
scikit-image>=0.19

# Monitoring
tensorboard>=2.10.0

# Optional: Streamlit comparison app
streamlit>=1.29.0
